+++ integration_tests/test_ibc.py
@@ -241,12 +241,6 @@ def test_cronos_transfer_source_tokens(ibc):
     with pytest.raises(AssertionError):
         cronos_cli.query_contract_by_denom(denom)
 
-    print("try token mapping with wrong denom, should fail")
-    rsp = cronos_cli.update_token_mapping(
-        denom, "0x000000000000000000000000000000000000dead", "DOG", 6, from_="validator"
-    )
-    assert rsp["code"] == 18, rsp["raw_log"]
-
     rsp = cronos_cli.update_token_mapping(
         denom, contract.address, "DOG", 6, from_="validator"
     )
@@ -300,6 +294,129 @@ def test_cronos_transfer_source_tokens(ibc):
     chainmain_receiver_balance = amount
     wait_for_fn("check balance change", check_chainmain_balance_change)
     assert chainmain_receiver_new_balance == amount + 1
+
+    # send back the token to cronos
+    # check receiver balance
+    cronos_balance_before_send = contract.caller.balanceOf(ADDRS["signer2"])
+    assert cronos_balance_before_send == 0
+
+    # send back token through ibc
+    print("Send back token through ibc")
+    chainmain_cli = ibc.chainmain.cosmos_cli()
+    cronos_receiver = eth_to_bech32(ADDRS["signer2"])
+
+    coin = "1000" + dest_denom
+    rsp = chainmain_cli.ibc_transfer(
+        chainmain_receiver, cronos_receiver, coin, "channel-0", 1, "100000000basecro"
+    )
+    assert rsp["code"] == 0, rsp["raw_log"]
+
+    # check contract balance
+    cronos_balance_after_send = 0
+
+    def check_contract_balance_change():
+        nonlocal cronos_balance_after_send
+        cronos_balance_after_send = contract.caller.balanceOf(ADDRS["signer2"])
+        return cronos_balance_after_send != cronos_balance_before_send
+
+    wait_for_fn("check contract balance change", check_contract_balance_change)
+    assert cronos_balance_after_send == amount
+
+
+def test_cronos_transfer_source_tokens_with_proxy(ibc):
+    """
+    test sending crc20 tokens originated from cronos to crypto-org-chain
+    """
+    assert_ready(ibc)
+    # deploy crc21 contract
+    w3 = ibc.cronos.w3
+    contract = deploy_contract(w3, CONTRACTS["TestCRC20"])
+
+    # setup the contract mapping
+    cronos_cli = ibc.cronos.cosmos_cli()
+
+    print("crc20 contract", contract.address)
+    denom = f"cronos{contract.address}"
+
+    print("check the contract mapping not exists yet")
+    with pytest.raises(AssertionError):
+        cronos_cli.query_contract_by_denom(denom)
+
+    rsp = cronos_cli.update_token_mapping(
+        denom, contract.address, "TEST", 6, from_="validator"
+    )
+    assert rsp["code"] == 0, rsp["raw_log"]
+    wait_for_new_blocks(cronos_cli, 1)
+
+    print("check the contract mapping exists now")
+    rsp = cronos_cli.query_denom_by_contract(contract.address)
+    assert rsp["denom"] == denom
+
+    # deploy crc20 proxy contract
+    proxycrc20 = deploy_contract(
+        w3,
+        CONTRACTS["TestCRC20Proxy"],
+        (contract.address, True),
+    )
+
+    print("proxycrc20 contract deployed at address: ", proxycrc20.address)
+    assert proxycrc20.caller.is_source()
+    assert proxycrc20.caller.crc20() == contract.address
+
+    # change token mapping
+    rsp = cronos_cli.update_token_mapping(
+        denom, proxycrc20.address, "TEST", 6, from_="validator"
+    )
+    assert rsp["code"] == 0, rsp["raw_log"]
+    wait_for_new_blocks(cronos_cli, 1)
+
+    print("check the contract mapping exists now")
+    rsp = cronos_cli.query_denom_by_contract(proxycrc20.address)
+    assert rsp["denom"] == denom
+
+    # send token to crypto.org
+    print("send to crypto.org")
+    chainmain_receiver = ibc.chainmain.cosmos_cli().address("signer2")
+    dest_denom = "ibc/67AC4BFF3CECEE5779A56B79301E0AECEACF21C531FF2979982970F6DFA25925"
+    amount = 1000
+    sender = ADDRS["validator"]
+
+    # First we need to approve the proxy contract to move asset
+    tx = contract.functions.approve(proxycrc20.address, amount).build_transaction(
+        {"from": sender}
+    )
+    txreceipt = send_transaction(w3, tx)
+    assert txreceipt.status == 1, "should success"
+    assert contract.caller.allowance(ADDRS["validator"], proxycrc20.address) == amount
+
+    # check and record receiver balance
+    chainmain_receiver_balance = get_balance(
+        ibc.chainmain, chainmain_receiver, dest_denom
+    )
+    assert chainmain_receiver_balance == 0
+
+    # send to ibc
+    tx = proxycrc20.functions.send_to_ibc(
+        chainmain_receiver, amount, 0, b""
+    ).build_transaction({"from": sender})
+    txreceipt = send_transaction(w3, tx)
+    print(txreceipt)
+    assert txreceipt.status == 1, "should success"
+
+    # check balance
+    chainmain_receiver_new_balance = 0
+
+    def check_chainmain_balance_change():
+        nonlocal chainmain_receiver_new_balance
+        chainmain_receiver_new_balance = get_balance(
+            ibc.chainmain, chainmain_receiver, dest_denom
+        )
+        chainmain_receiver_all_balance = get_balances(ibc.chainmain, chainmain_receiver)
+        print("receiver all balance:", chainmain_receiver_all_balance)
+        return chainmain_receiver_balance != chainmain_receiver_new_balance
+
+    wait_for_fn("check balance change", check_chainmain_balance_change)
+    assert chainmain_receiver_new_balance == amount
 
     # send back the token to cronos
     # check receiver balance